first batch of x87 instructions
added sti aliases for FPU registers (rather than just st and st(i)).

added unified formatters for binary (i.e. of form a <- f(a, b) ) fpu instructions.
added the FADD family.
added the FSUB/FSUBR families.
added the FMUL family.
added the FDIV/FDIVR families.

marked TryGet_cc() as private (was never made private since the merge from Utility class).

fixed a bug where PushFPU() and PopFPU() weren't setting the InUse flags correctly.
fixed a bug where TOP setter was storing a byte instead of a word.

added FPUAccessViolation error for using a register that's not in use

+++ src/CSX64/Registers.cs
@@ -201,7 +201,7 @@ namespace CSX64
         public byte TOP
         {
             get => (byte)((FPU_status & 0x3800) >> 11);
-            set => FPU_status = (byte)(FPU_status & ~0x3800 | ((value & 7) << 11));
+            set => FPU_status = (UInt16)(FPU_status & ~0x3800 | ((value & 7) << 11));
         }
 
         public UInt64 ST0 { get => FPURegisters[(TOP + 0) & 7].Value; set => FPURegisters[(TOP + 0) & 7].Value = value; }
